Refactoring InputManager and Grid for clarity and bug fixing
The InputManager class now has clearer method names for keypress handling. For instance, `HandleTKeyPress()` is now `HandleTheMirrorCopyButtonT()`, reflecting its specific functionality. Grid ZoomIn and ZoomOut methods now include checks and casts to assure images don't break when zooming beyond images capacity. Grid.cs was updated with BoxSize changing from int to float to fix type mismatch bugs. The handle method for the mirror copy button in the InputManager class was also enhanced to include more specific start and end coordinates, improving mapping accuracy.

Refactoring MatrixManipulation for clearer output

The `MirrorMatrix` method in the MatrixManipulation doesn't return a boolean array anymore. Now, it returns the same type as the input (Cell[,]). A new method `MakeBoolGrid` was created to handle the transformation from the Cell array to boolean array. The method now returns an array of bool basis of cell's isAlive property.

+++ Grid.cs
@@ -20,7 +20,7 @@ public partial class Grid: Node2D
 	public float UpdateTickRate { get; set; } = 0.5f;
 	public bool ResetHigherGridArray { get; set; }
 	public List<Cell[,]> ListOfCellArrayStates { get; set; } = new(); 
-	public int BoxSize { get; set; } = 10;
+	public float BoxSize { get; set; } = 10;
 	public bool DebugState { get; set; } = true;
 	public bool DrawCopyBox { get; set; } = false;
 
@@ -47,8 +47,8 @@ public partial class Grid: Node2D
 	private void InitGrid()
 	{
 		var viewportSize = _gridSize;;
-		GridWidth = (int)viewportSize.X / BoxSize;
-		GridHeight = (int)viewportSize.Y / BoxSize;
+		GridWidth = (int)viewportSize.X / (int)BoxSize;
+		GridHeight = (int)viewportSize.Y / (int)BoxSize;
 		var emptyCellGrid = PatternCreator.GetInitialGrid(GridWidth, GridHeight);
 
 		var initialCellGrid = PatternCreator.CreatePattern(PatternCreator.Pattern.Star, emptyCellGrid);
